dom/html/test/forms/without_selectionchange/test_select.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Test for Bug 1717435</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>

<textarea id="textarea">foo</textarea>
<script>
  SimpleTest.waitForExplicitFinish();

  textarea.addEventListener("select", ev => {
    ok(true, "A select event must fire regardless of dom.select_events.textcontrols.enabled");
    SimpleTest.finish();
  });

  textarea.focus();
  textarea.select();
  is(textarea.selectionStart, 0, "selectionStart")
  is(textarea.selectionEnd, 3, "selectionEnd")
</script>
